hybrid: bind ssh dir from controller to hypervisor
The hypervisor container launches a custom ssh server listening on an
alternate port by default. When the controller goes to launch a VM it
performs a scp of the virsh domain between controller and hypervisor via
a datastore dir in /var/lib/one. This scp looks like:

```
$ scp $CONTROLLER_HOSTNAME:/path $HYPERVISOR_HOSTNAME:/path
```

This means the controller first ssh's to itself and then performs the
scp from the local /var/lib/one directory to the hypervisor host. All
corona containers ssh daemons and client configs are configured to use
an alternate default port reserved for opennebula/corona communication.
In the hybrid case the hypervisor's custom ssh runs on the same host as
the controller which means the above scp first ssh's inside of the
hypervisor container and then does a scp from hypervisor to itself. In
order for this to work it needs to share and authorize the same ssh
credentials and config as the controller otherwise ssh will land on the
controller hosts "system" ssh running on port 22 and prompts for a
password.