libvirt: ensure KVM gid matches /dev/kvm gid
This is needed to launch VMs as oneadmin. The oneadmin user must belong
to a 'kvm' group that has R/W permission to /dev/kvm. In order to ensure
this is the case we check that /dev/kvm is group R/W (and error out
otherwise) and update the container's 'kvm' group ID to match /dev/kvm's
GID if necessary. Both the kvm check and GID update can be disabled via
the following environment variables:

$KVM_SANITY_CHECK
$KVM_SETUP

Both operations are enabled by default.