controller: clean up stale oned lock files
These locks get left around due to the fact that supervisord doesn't
support program "dependencies" which means we have to set each program
with autostart=false, run supervisord detached in the background instead
of in the foreground, and then call supervisorctl in serial for each
program in the required order. Then in order to keep the container
running we tail the supervisor log file indefinitely which effectively
makes tail the "main"/"init" process. When docker stop sends its signal
the tail program gets SIGTERM and exits while the detached supervisord
(and its programs) in the background gets forcibly killed leaving a
stale lock file.

It might be possible to solve this by using supervisor's event listeners
to handle the start-up order problem which would allow us to enable
autostart and keep supervisord in the foreground. In this case
supervisord would actually receive the SIGTERM signal and do an orderly
shutdown of each process which would clean up the opennebula lock file
properly. For now just clean up any stale lock files on container
start-up.