base: add supervisord launch script
This script handles launching supervisord in the foreground and,
provided $SUPERVISORD_ORDERED="yes" in the environment, will parse all
programs in the /etc/supervisor.d/*.conf files and launch them in serial
in the order it finds them. It uses standard shell globbing when
ordering the files for parsing and then parses the [program:xxx]
sections in the files from top to bottom.

For example:

/etc/supervisor.d/00-first.conf
  [program:foo]
  [program:bar]
/etc/supervisor.d/05-second.conf
  [program:kong]
  [program:donkey]
/etc/supervisor.d/10-third.conf
  [program:zoo]

In this case the above programs would be started in serial in the
following order: foo, bar, kong, donkey, zoo.

This script serves as a utility for derivative container images.